fix bug passing extra to scoreboard activity from main menu

+++ app/src/main/java/com/hit/game012/MenuActivity.java
@@ -52,7 +52,7 @@ public class MenuActivity extends AppCompatActivity {
         intent.putExtra("boardSize", 4);
         intent.putExtra("isDailyGame", true);
         intent.putExtra("win", false);
-        intent.putExtra("gameTime", (long)0);
+        intent.putExtra("gameTime", (int)0);
         intent.putExtra("hintCounter", 0);
         startActivity(intent);
     }
